backend/requirements.txt
Django==5.1.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==12.1.1
python-decouple==3.8
cryptography==46.0.5
reportlab==4.2.5
qrcode[pil]==7.4.2
requests==2.32.3
mammoth==1.8.0
pypdf==5.1.0
pyotp==2.9.0
gunicorn==23.0.0
whitenoise==6.8.2
